Check if song list was successfully fetched.

+++ src/usdb_syncer/song_routines.py
@@ -30,7 +30,7 @@ from usdb_syncer.usdb_song import UsdbSong, UsdbSongEncoder
 from usdb_syncer.utils import AppPaths
 
 
-def load_available_songs(force_reload: bool, session: Session | None = None) -> None:
+def load_available_songs(force_reload: bool, session: Session | None = None) -> bool:
     if force_reload:
         max_skip_id = SongId(0)
         UsdbSong.delete_all()
@@ -41,14 +41,15 @@ def load_available_songs(force_reload: bool, session: Session | None = None) -> 
         songs = get_usdb_available_songs(max_skip_id, session=session)
     except errors.UsdbLoginError:
         logger.debug("Skipping fetching new songs as there is no login.")
-        return
+        return False
     except requests.exceptions.ConnectionError:
         logger.debug("", exc_info=True)
         logger.error("Failed to fetch new songs; check network connection.")
-        return
+        return False
     if songs:
         UsdbSong.upsert_many(songs)
         _download_subscribed_songs(songs)
+    return True
 
 
 def _download_subscribed_songs(songs: list[UsdbSong]) -> None:
@@ -95,7 +96,7 @@ def _iterate_usdb_files_in_folder_recursively(
                 yield Path(root) / file
 
 
-def synchronize_sync_meta_folder(folder: Path) -> None:
+def synchronize_sync_meta_folder(folder: Path, song_list_updated: bool) -> None:
     db_metas = {m.sync_meta_id: m for m in SyncMeta.get_in_folder(folder)}
     song_ids = set(db.all_song_ids())
     to_upsert: list[SyncMeta] = []
@@ -128,12 +129,13 @@ def synchronize_sync_meta_folder(folder: Path) -> None:
                 else:
                     logger.info(f"New meta file found on disk: '{path}'.")
             else:
-                logger.info(
-                    f"{meta.song_id.usdb_detail_url()} no longer exists on USDB."
-                )
-                if settings.get_trash_remotely_deleted_songs():
-                    logger.info(f"Deleting '{path.parent}' locally as well.")
-                    send2trash.send2trash(path.parent)
+                if song_list_updated:
+                    logger.info(
+                        f"{meta.song_id.usdb_detail_url()} no longer exists on USDB."
+                    )
+                    if settings.get_trash_remotely_deleted_songs():
+                        logger.info(f"Deleting '{path.parent}' locally as well.")
+                        send2trash.send2trash(path.parent)
 
     SyncMeta.delete_many(tuple(db_metas.keys() - found_metas))
     SyncMeta.upsert_many(to_upsert)
